requirements.txt
aiohttp==3.7.4.post0
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
discord.py==1.6.0
finnhub-python==2.4.0
giphy-client==1.0.0
idna==2.8
isort==4.3.21
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.4
mccabe==0.6.1
multidict==4.5.2
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2021.1
requests==2.25.1
six==1.13.0
typing-extensions==3.7.4.3
urllib3==1.25.7
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
